add form validation on server
